+++ src/ast/nodes/ClassBody.ts
@@ -2,8 +2,8 @@ import { NodeBase } from './shared/Node';
 import ExecutionPathOptions from '../ExecutionPathOptions';
 import CallOptions from '../CallOptions';
 import MethodDefinition from './MethodDefinition';
-import { ObjectPath } from '../variables/VariableReassignmentTracker';
 import { NodeType } from './NodeType';
+import { ObjectPath } from '../values';
 
 export default class ClassBody extends NodeBase {
 	type: NodeType.ClassBody;
