+++ src/ast/nodes/ClassExpression.ts
@@ -1,8 +1,8 @@
 import ClassNode from './shared/ClassNode';
 import Scope from '../scopes/Scope';
-import { ObjectPath } from '../variables/VariableReassignmentTracker';
 import ExecutionPathOptions from '../ExecutionPathOptions';
 import { NodeType } from './NodeType';
+import { ObjectPath } from '../values';
 
 export default class ClassExpression extends ClassNode {
 	type: NodeType.ClassExpression;
