+++ src/ast/nodes/ExpressionStatement.ts
@@ -1,7 +1,7 @@
 import MagicString from 'magic-string';
 import Scope from '../scopes/Scope';
-import { RenderOptions } from '../../Module';
 import { StatementBase } from './shared/Node';
+import { RenderOptions } from '../../utils/renderHelpers';
 
 export default class ExpressionStatement extends StatementBase {
 	directive?: string;
