+++ src/ast/nodes/Import.ts
@@ -3,7 +3,7 @@ import { NodeType } from './NodeType';
 import { NodeBase } from './shared/Node';
 import MagicString from 'magic-string';
 import NamespaceVariable from '../variables/NamespaceVariable';
-import { RenderOptions } from '../../Module';
+import { RenderOptions } from '../../utils/renderHelpers';
 
 export default class Import extends NodeBase {
 	type: NodeType.Import;
