+++ src/ast/nodes/ObjectPattern.ts
@@ -1,11 +1,11 @@
 import AssignmentProperty from './AssignmentProperty';
 import Scope from '../scopes/Scope';
 import ExecutionPathOptions from '../ExecutionPathOptions';
-import { ObjectPath } from '../variables/VariableReassignmentTracker';
 import { ExpressionEntity } from './shared/Expression';
 import { PatternNode } from './shared/Pattern';
 import { NodeBase } from './shared/Node';
 import { NodeType } from './NodeType';
+import { ObjectPath } from '../values';
 
 export default class ObjectPattern extends NodeBase implements PatternNode {
 	type: NodeType.ObjectPattern;
