+++ src/ast/nodes/Program.ts
@@ -1,8 +1,7 @@
 import MagicString from 'magic-string';
 import { NodeBase, StatementNode } from './shared/Node';
 import { NodeType } from './NodeType';
-import { RenderOptions } from '../../Module';
-import { renderStatementList } from '../../utils/renderHelpers';
+import { RenderOptions, renderStatementList } from '../../utils/renderHelpers';
 
 export default class Program extends NodeBase {
 	type: NodeType.Program;
