+++ src/ast/nodes/SequenceExpression.ts
@@ -2,7 +2,7 @@ import ExecutionPathOptions from '../ExecutionPathOptions';
 import MagicString from 'magic-string';
 import { ExpressionNode, NodeBase } from './shared/Node';
 import { NodeType } from './NodeType';
-import { RenderOptions } from '../../Module';
+import { RenderOptions } from '../../utils/renderHelpers';
 
 export default class SequenceExpression extends NodeBase {
 	type: NodeType.SequenceExpression;
