+++ src/ast/nodes/UnaryExpression.ts
@@ -1,6 +1,5 @@
-import { UNKNOWN_VALUE } from '../values';
+import { ObjectPath, UNKNOWN_VALUE } from '../values';
 import ExecutionPathOptions from '../ExecutionPathOptions';
-import { ObjectPath } from '../variables/VariableReassignmentTracker';
 import { NodeType } from './NodeType';
 import { ExpressionNode, NodeBase } from './shared/Node';
